// Dashboard/src/Types/Tables.type.ts
export interface School {
    _id: string;
    schoolname: string;
    userid: { email: string };
    address: string;
    contactno: string;
    password: string;
    file: string;
}